+++ elide-standalone/src/main/java/com/yahoo/elide/standalone/ElideStandalone.java
@@ -1,0 +1,98 @@
+/*
+ * Copyright 2017, Oath Inc.
+ * Licensed under the Apache License, Version 2.0
+ * See LICENSE file in project root for terms.
+ */
+package com.yahoo.elide.standalone;
+
+import com.yahoo.elide.resources.DefaultOpaqueUserFunction;
+import com.yahoo.elide.security.checks.Check;
+import com.yahoo.elide.standalone.config.ElideResourceConfig;
+import com.yahoo.elide.standalone.config.ElideStandaloneSettings;
+import lombok.extern.slf4j.Slf4j;
+import org.eclipse.jetty.server.Server;
+import org.eclipse.jetty.servlet.ServletContextHandler;
+import org.eclipse.jetty.servlet.ServletHolder;
+import org.glassfish.jersey.servlet.ServletContainer;
+
+import javax.ws.rs.core.SecurityContext;
+import java.util.Map;
+
+import static com.yahoo.elide.standalone.config.ElideResourceConfig.ELIDE_STANDALONE_SETTINGS_ATTR;
+
+@Slf4j
+public class ElideStandalone {
+    private final ElideStandaloneSettings elideStandaloneSettings;
+
+    /**
+     * Constructor
+     *
+     * @param elideStandaloneSettings Elide standalone configuration settings.
+     */
+    public ElideStandalone(ElideStandaloneSettings elideStandaloneSettings) {
+        this.elideStandaloneSettings = elideStandaloneSettings;
+    }
+
+    /**
+     * Constructor
+     *
+     * @param checkMappings Check mappings to use for service.
+     */
+    public ElideStandalone(Map<String, Class<? extends Check>> checkMappings) {
+        this(checkMappings, SecurityContext::getUserPrincipal);
+    }
+
+    /**
+     * Constructor
+     *
+     * @param checkMappings Check mappings to use for service.
+     * @param userExtractionFn User extraction function to use for service.
+     */
+    public ElideStandalone(Map<String, Class<? extends Check>> checkMappings,
+                           DefaultOpaqueUserFunction userExtractionFn) {
+        this(new ElideStandaloneSettings() {
+            @Override
+            public Map<String, Class<? extends Check>> getCheckMappings() {
+                return checkMappings;
+            }
+
+            @Override
+            public DefaultOpaqueUserFunction getUserExtractionFunction() {
+                return userExtractionFn;
+            }
+        });
+    }
+
+    /**
+     * Start the Elide service.
+     *
+     * This method blocks.
+     */
+    public void start() throws Exception {
+        ServletContextHandler context = new ServletContextHandler();
+        context.setContextPath("/");
+
+        log.info("Starting jetty server on port: {}", elideStandaloneSettings.getPort());
+        Server jettyServer = new Server(elideStandaloneSettings.getPort());
+        jettyServer.setHandler(context);
+
+        context.setAttribute(ELIDE_STANDALONE_SETTINGS_ATTR, elideStandaloneSettings);
+
+        ServletHolder jerseyServlet = context.addServlet(ServletContainer.class,
+                elideStandaloneSettings.getJsonApiPathSpec());
+        jerseyServlet.setInitOrder(0);
+        jerseyServlet.setInitParameter("jersey.config.server.provider.packages", "com.yahoo.elide.resources");
+        jerseyServlet.setInitParameter("javax.ws.rs.Application", ElideResourceConfig.class.getCanonicalName());
+
+        try {
+            jettyServer.start();
+            log.info("Jetty started!");
+            jettyServer.join();
+        } catch (Exception e) {
+            log.error("Unexpected exception caught: {}", e.getMessage(), e);
+            throw e;
+        } finally {
+            jettyServer.destroy();
+        }
+    }
+}